pyproject.toml
[tool.poetry]
name = "matomo4-plugins"
version = "0.1.0"
description = ""
authors = ["Lukas Winkler <git@lw1.at>"]

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.0.0"
requests = "^2.24.0"
pyyaml = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
